xcarchive -> ipa
