+++ packages/flutter_tools/lib/src/application_package.dart
@@ -386,9 +386,11 @@ class BuildableIOSApp extends IOSApp {
     => globals.fs.path.join(getIosBuildDirectory(), 'archive', globals.fs.path.withoutExtension(_hostAppBundleName));
 
   // The output xcarchive bundle path `build/ios/archive/Runner.xcarchive`.
-  String get archiveBundleOutputPath => globals.fs.path.setExtension(archiveBundlePath, '.xcarchive');
+  String get archiveBundleOutputPath =>
+      globals.fs.path.setExtension(archiveBundlePath, '.xcarchive');
 
-  String get ipaOutputPath => globals.fs.path.join(getIosBuildDirectory(), 'ipa');
+  String get ipaOutputPath =>
+      globals.fs.path.join(getIosBuildDirectory(), 'ipa');
 
   String _buildAppPath(String type) {
     return globals.fs.path.join(getIosBuildDirectory(), type, _hostAppBundleName);
