+++ packages/flutter_tools/lib/src/commands/build.dart
@@ -15,7 +15,6 @@ import 'build_bundle.dart';
 import 'build_fuchsia.dart';
 import 'build_ios.dart';
 import 'build_ios_framework.dart';
-import 'build_ipa.dart';
 import 'build_web.dart';
 
 class BuildCommand extends FlutterCommand {
@@ -30,7 +29,6 @@ class BuildCommand extends FlutterCommand {
       verboseHelp: verboseHelp,
     ));
     addSubcommand(BuildIOSArchiveCommand(verboseHelp: verboseHelp));
-    addSubcommand(BuildIPACommand());
     addSubcommand(BuildBundleCommand(verboseHelp: verboseHelp));
     addSubcommand(BuildWebCommand(verboseHelp: verboseHelp));
     addSubcommand(BuildMacosCommand(verboseHelp: verboseHelp));
